linux autostart wip
