+++ src/qz/deploy/LinuxDeploy.java
@@ -11,7 +11,13 @@
 package qz.deploy;
 
 import java.awt.Toolkit;
+import java.io.IOException;
 import java.lang.reflect.Field;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.nio.file.StandardOpenOption;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import qz.common.Constants;
@@ -24,54 +30,13 @@ class LinuxDeploy extends DeployUtilities {
 
     private static final Logger log = LoggerFactory.getLogger(LinuxDeploy.class);
 
-    private static String STARTUP = "/etc/xdg/autostart/";
     private static String DESKTOP = System.getProperty("user.home") + "/Desktop/";
-    private static String DELETED_ICON = "process-stop";
 
     private String appLauncher = "/usr/share/applications/" + getShortcutName();
-
-    //@Override
-    //public boolean createStartupShortcut() {
-    //    return copyShortcut(appLauncher, STARTUP);
-    //}
 
     @Override
     public boolean createDesktopShortcut() {
         return copyShortcut(appLauncher, DESKTOP);
-    }
-
-    //@Override
-    //public boolean removeStartupShortcut() {
-    //    return writeArrayToFile(STARTUP + getShortcutName(), new String[]{
-    //            "[Desktop Entry]",
-    //            "Type=Application",
-    //            "Name=" + Constants.ABOUT_TITLE + " (Disabled)",
-    //            "Exec=/bin/true",
-    //            "Icon=" + DELETED_ICON,
-    //            "Terminal=false",
-    //    });
-    //}
-
-
-    //@Override
-    //public boolean hasStartupShortcut() {
-    //    try {
-    //        String contents = FileUtilities.readLocalFile(STARTUP + getShortcutName());
-    //        return !contents.contains("Icon=" + DELETED_ICON);
-    //    } catch (Exception ex) {
-    //        log.warn("Unable to read startup shortcut {}", STARTUP + getShortcutName(), ex);
-    //    }
-    //    return false;
-    //}
-
-    @Override
-    public boolean setAutostart(boolean autostart) {
-        return false;
-    }
-
-    @Override
-    public boolean isAutostart() {
-        return false;
     }
 
     private static boolean copyShortcut(String source, String target) {
